+++ CHANGES.md
@@ -6,6 +6,11 @@
 
  * Move markup related data types to haddock-library
 
+ * Fix: Show empty constraint contexts in pattern type signatures (#663)
+
+ * Fix: Generate constraint signatures for constructors exported as pattern
+   synonyms (#663)
+
 ## Changes in version 2.18.1
 
  * Synopsis is working again (#599)